refactor(useProfile): separate concerns into domain-specific hooks that can be individually mantained and tested -
